Tailwind CSSとnext/fontを適応
