+++ apps/next-app/postcss.config.mjs
@@ -1,5 +1,6 @@
-const config = {
-	plugins: ['@tailwindcss/postcss'],
+/** @type {import('tailwindcss').Config} */
+export default {
+	plugins: {
+		'@tailwindcss/postcss': {},
+	},
 };
-
-export default config;
